refactor: 이미지 업로드 로직 useImageUpload 훅으로 분리
